+++ src/core/data.cpp
@@ -61,6 +61,13 @@ void DataManager::writeTTSConfig(const TTSConfig &ttsc)
     json_object.insert("tr_lang_translators", ttsc.tr_lang_translators);
     json_object.insert("tr_lang_libretranslate", ttsc.tr_lang_libretranslate);
 
+    json_object.insert("tr_tx_secret_id", ttsc.tr_tx_secret_id);
+    json_object.insert("tr_tx_secret_key", ttsc.tr_tx_secret_key);
+    json_object.insert("tr_tx_region", ttsc.tr_tx_region);
+    json_object.insert("tr_tx_project_id", ttsc.tr_tx_project_id);
+    json_object.insert("tr_tx_source_lang", ttsc.tr_tx_source_lang);
+    json_object.insert("tr_tx_target_lang", ttsc.tr_tx_target_lang);
+
     QFile file(TTS_CONFIG_FILE);
     if (!file.open(QIODevice::WriteOnly))
     {
@@ -254,6 +261,13 @@ void DataManager::readTTSConfig()
     tts_config.tr_libretranslate_port = json_object.value("tr_libretranslate_port").toString("");
     tts_config.tr_provider = json_object.value("tr_provider").toString("");
     tts_config.tr_lang_translators = json_object.value("tr_lang_translators").toString("");
+    // 腾讯云机器翻译配置 tx_tr
+    tts_config.tr_tx_secret_id = json_object.value("tr_tx_secret_id").toString("");
+    tts_config.tr_tx_secret_key = json_object.value("tr_tx_secret_key").toString("");
+    tts_config.tr_tx_region = json_object.value("tr_tx_region").toString("");
+    tts_config.tr_tx_project_id = json_object.value("tr_tx_project_id").toInt(0);
+    tts_config.tr_tx_source_lang = json_object.value("tr_tx_source_lang").toString("auto");
+    tts_config.tr_tx_target_lang = json_object.value("tr_tx_target_lang").toString("");
 }
 void DataManager::readOpenWeatherData()
 {